全部aio，mysql上bug处理

+++ app/models/asset_snapshot.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime, Numeric, Index
+from sqlalchemy import Column, Integer, String, DateTime, Numeric, Index, BigInteger
 from datetime import datetime, UTC
 from app.models.base import BaseModel
 
 class AssetSnapshot(BaseModel):
     __tablename__ = "asset_snapshots"
 
-    project_id = Column(Integer, nullable=False, index=True, comment="项目ID")
+    project_id = Column(BigInteger, nullable=False, index=True, comment="项目ID")
     snapshot_time = Column(DateTime, nullable=False, index=True, comment="快照时间")
     
     # 资产相关字段
