+++ app/models/position.py
@@ -1,13 +1,13 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, Numeric, JSON
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, Numeric, JSON, BigInteger
 from datetime import datetime, UTC
 from app.models.base import BaseModel
 
 class Position(BaseModel):
     __tablename__ = "positions"
 
-    strategy_id = Column(Integer, index=True, nullable=False, comment="策略ID")
+    strategy_id = Column(BigInteger, index=True, nullable=False, comment="策略ID")
     strategy_instance_id = Column(String(200), index=True, nullable=False, comment="策略实例ID")
-    project_id = Column(Integer, nullable=False,index=True)
+    project_id = Column(BigInteger, nullable=False,index=True)
     symbol = Column(String(32), nullable=False, comment="交易标的")
     quote_currency = Column(String(16), nullable=False, comment="计价货币")
     ins_type = Column(String(16), nullable=False, comment="合约/现货类型")
