+++ app/models/signal.py
@@ -1,15 +1,15 @@
-from sqlalchemy import Column, String, DateTime, JSON, Integer
+from sqlalchemy import Column, String, DateTime, JSON, Integer, BigInteger
 
 from app.models.base import Base, BaseModel
 
 class Signal(Base):
     __tablename__ = "signals"
 
-    id = Column(Integer, primary_key=True, index=True)
-    project_id = Column(Integer, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
+    project_id = Column(BigInteger, index=True)
     data_source_instance_id =  Column(JSON)
     data_source_class_name = Column(String(200), nullable=False)
-    strategy_id =  Column(Integer, index=True)
+    strategy_id =  Column(BigInteger, index=True)
     strategy_instance_id =  Column(String(200), nullable=False)
     strategy_class_name = Column(String(200), nullable=False)
     signal_time = Column(DateTime, index=True, nullable=False)
